requirements.txt
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
fastapi==0.110.0
grpcio==1.80.0
grpcio-tools==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
numpy==2.4.4
openai==1.30.1
portalocker==2.10.1
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
qdrant-client==1.9.0
redis==7.4.0
regex==2026.3.32
requests==2.33.1
sniffio==1.3.1
starlette==0.36.3
tiktoken==0.7.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
